streamlit/requirements.txt
streamlit==1.32.2
requests==2.31.0
Pillow==10.2.0